On startup change to PrePrepare/Normal in certain states
If using the persistent PBFT state log, it is possible to
stop and restart the engine in a state where it is expecting
a response.  Upon the restart the engine will resume that
state, but given sufficient downtime, those responses will
never come.  In those cases we should reset to
PrePreparing/Normal and allow the algorithm to sort things out.
The problem states are ViewChanging, and Finishing(true)

Signed-off-by: Kevin O'Donnell <kevin@blockchaintp.com>